+++ src/main/java/com/exaroton/api/APIRequest.java
@@ -1,6 +1,5 @@
 package com.exaroton.api;
 
-
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;