+++ src/main/java/com/exaroton/api/request/server/GetServersRequest.java
@@ -1,6 +1,5 @@
 package com.exaroton.api.request.server;
 
-
 import com.exaroton.api.APIRequest;
 import com.exaroton.api.APIResponse;
 import com.exaroton.api.server.Server;