+++ src/main/java/com/exaroton/api/server/ServerLog.java
@@ -1,14 +1,8 @@
 package com.exaroton.api.server;
 
-import org.jetbrains.annotations.ApiStatus;
-
+@SuppressWarnings("unused")
 public final class ServerLog {
-    private final String content;
-
-    @ApiStatus.Internal
-    public ServerLog(String content) {
-        this.content = content;
-    }
+    private String content;
 
     public String getContent() {
         return content;