+++ src/test/java/APIClientTest.java
@@ -6,4 +6,12 @@ public abstract class APIClientTest {
     protected static final @NotNull String TEST_SERVER_ID = System.getenv("EXAROTON_TEST_SERVER");
     protected final @NotNull ExarotonClient client = new ExarotonClient(System.getenv("EXAROTON_API_TOKEN"));
     protected final @NotNull Server server = client.getServer(TEST_SERVER_ID);
+
+    protected APIClientTest() {
+        try {
+            Thread.sleep(100);
+        } catch (InterruptedException e) {
+            throw new RuntimeException(e);
+        }
+    }
 }