+++ src/test/java/FileTest.java
@@ -16,7 +16,7 @@ import static org.junit.jupiter.api.Assertions.*;
 
 public class FileTest extends APIClientTest {
     @Test
-    void getFile() throws APIException, IOException {
+    void getFile() throws IOException {
         ServerFile whitelist = server.getFile("whitelist.json");
         whitelist.putContent("[{\"name\":\"JulianVennen\", \"uuid\": \"abcd9e56-5ac2-490c-8bc9-6c1cad18f506\"}]");
         assertNotNull(whitelist);